+++ driver_hydrocalm3.h
@@ -13,8 +13,8 @@
 struct Hydrocalm3: Driver
 {
   Hydrocalm3(std::string key = "") : Driver(std::string("hydrocalm3"), key) {};
-  virtual esphome::optional<std::map<std::string, float>> get_values(std::vector<unsigned char> &telegram) override {
-    std::map<std::string, float> ret_val{};
+  virtual esphome::optional<std::map<std::string, double>> get_values(std::vector<unsigned char> &telegram) override {
+    std::map<std::string, double> ret_val{};
 
     add_to_map(ret_val, "total_heating_kwh", this->get_0C0E(telegram));
     add_to_map(ret_val, "total_heating_kwh", this->get_0C03(telegram));
