+++ driver_mkradio3.h
@@ -13,8 +13,8 @@
 struct Mkradio3: Driver
 {
   Mkradio3(std::string key = "") : Driver(std::string("mkradio3"), key) {};
-  virtual esphome::optional<std::map<std::string, float>> get_values(std::vector<unsigned char> &telegram) override {
-    std::map<std::string, float> ret_val{};
+  virtual esphome::optional<std::map<std::string, double>> get_values(std::vector<unsigned char> &telegram) override {
+    std::map<std::string, double> ret_val{};
 
     add_to_map(ret_val, "total_water_m3", this->get_total_water_m3(telegram));
 
@@ -27,8 +27,8 @@ struct Mkradio3: Driver
   };
 
 private:
-  esphome::optional<float> get_total_water_m3(std::vector<unsigned char> &telegram) {
-    esphome::optional<float> ret_val{};
+  esphome::optional<double> get_total_water_m3(std::vector<unsigned char> &telegram) {
+    esphome::optional<double> ret_val{};
     size_t i = 11;
 
     ret_val = ((((uint32_t)telegram[i+4] << 8) + (uint32_t)telegram[i+3]) / 10.0) +
